+++ src/main/java/com/baomidou/dynamic/datasource/DynamicDataSourceCreator.java
@@ -1,3 +1,19 @@
+/**
+ * Copyright © 2018 organization baomidou
+ * <pre>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * <pre/>
+ */
 package com.baomidou.dynamic.datasource;
 
 import com.alibaba.druid.pool.DruidDataSource;
@@ -16,7 +32,7 @@ import java.sql.SQLException;
  * @author TaoYu
  * @since 2.3.0
  */
-public class DataSourceCreator {
+public class DynamicDataSourceCreator {
 
     private static final String DRUID_DATASOURCE = "com.alibaba.druid.pool.DruidDataSource";
 
@@ -28,7 +44,7 @@ public class DataSourceCreator {
     private Method driverClassNameMethod;
     private Method buildMethod;
 
-    public DataSourceCreator() {
+    public DynamicDataSourceCreator() {
         Class builderClass = null;
         try {
             builderClass = Class.forName("org.springframework.boot.jdbc.DataSourceBuilder");