+++ src/main/java/com/baomidou/dynamic/datasource/provider/AbstractJdbcDataSourceProvider.java
@@ -16,7 +16,7 @@
  */
 package com.baomidou.dynamic.datasource.provider;
 
-import com.baomidou.dynamic.datasource.DataSourceCreator;
+import com.baomidou.dynamic.datasource.DynamicDataSourceCreator;
 import com.baomidou.dynamic.datasource.spring.boot.autoconfigure.DataSourceProperty;
 import com.baomidou.dynamic.datasource.spring.boot.autoconfigure.DynamicDataSourceProperties;
 import lombok.extern.slf4j.Slf4j;
@@ -44,7 +44,7 @@ public abstract class AbstractJdbcDataSourceProvider implements DynamicDataSourc
     protected DynamicDataSourceProperties dynamicDataSourceProperties;
 
     @Autowired
-    private DataSourceCreator dataSourceCreator;
+    private DynamicDataSourceCreator dynamicDataSourceCreator;
 
     /**
      * JDBC driver
@@ -86,7 +86,7 @@ public abstract class AbstractJdbcDataSourceProvider implements DynamicDataSourc
             Map<String, DataSourceProperty> dataSourcePropertiesMap = executeStmt(stmt);
             Map<String, DataSource> dataSourceMap = new HashMap<>(dataSourcePropertiesMap.size());
             for (Map.Entry<String, DataSourceProperty> item : dataSourcePropertiesMap.entrySet()) {
-                dataSourceMap.put(item.getKey(), dataSourceCreator.createDataSource(item.getValue(), dynamicDataSourceProperties.getDruid()));
+                dataSourceMap.put(item.getKey(), dynamicDataSourceCreator.createDataSource(item.getValue(), dynamicDataSourceProperties.getDruid()));
             }
             return dataSourceMap;
         } catch (Exception e) {